Moved files to its correct folder
