+++ EventAppTest2/app/src/main/java/com/example/eventapptest2/OpenEventsRecyclerViewAdapter.java
