+++ EventAppTest2/app/src/main/java/com/example/eventapptest2/OrgainzersEventFragment.java
@@ -1,12 +1,18 @@
 package com.example.eventapptest2;
 
+import android.content.Context;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.eventapptest2.placeholder.PlaceholderContent;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -17,10 +23,12 @@ public class OrgainzersEventFragment extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+    private static final String ARG_COLUMN_COUNT = "column-count";
 //    private static final String ARG_PARAM1 = "param1";
 //    private static final String ARG_PARAM2 = "param2";
 //
 //    // TODO: Rename and change types of parameters
+    private  int mColumnCount = 1;
 //    private String mParam1;
 //    private String mParam2;
 
@@ -32,33 +40,44 @@ public class OrgainzersEventFragment extends Fragment {
      * Use this factory method to create a new instance of
      * this fragment using the provided parameters.
      *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
      * @return A new instance of fragment OrgainzersEventFragment.
      */
     // TODO: Rename and change types and number of parameters
-//    public static OrgainzersEventFragment newInstance(String param1, String param2) {
-//        OrgainzersEventFragment fragment = new OrgainzersEventFragment();
-//        Bundle args = new Bundle();
-//        args.putString(ARG_PARAM1, param1);
-//        args.putString(ARG_PARAM2, param2);
-//        fragment.setArguments(args);
-//        return fragment;
-//    }
+    public static OrgainzersEventFragment newInstance(int columnCount) {
+        OrgainzersEventFragment fragment = new OrgainzersEventFragment();
+        Bundle args = new Bundle();
+        args.putInt(ARG_COLUMN_COUNT, columnCount);
+        fragment.setArguments(args);
+        return fragment;
+    }
 //
-//    @Override
-//    public void onCreate(Bundle savedInstanceState) {
-//        super.onCreate(savedInstanceState);
-//        if (getArguments() != null) {
-//            mParam1 = getArguments().getString(ARG_PARAM1);
-//            mParam2 = getArguments().getString(ARG_PARAM2);
-//        }
-//    }
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        if (getArguments() != null) {
+            mColumnCount = getArguments().getInt(ARG_COLUMN_COUNT);
+            //mParam2 = getArguments().getString(ARG_PARAM2);
+        }
+    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_orgainzers_event, container, false);
+        View view = inflater.inflate(R.layout.open_events_list, container, false);
+
+        // Set the adapter
+        if (view instanceof RecyclerView) {
+            Context context = view.getContext();
+            RecyclerView recyclerView = (RecyclerView) view;
+            if (mColumnCount <= 1) {
+                recyclerView.setLayoutManager(new LinearLayoutManager(context));
+            } else {
+                recyclerView.setLayoutManager(new GridLayoutManager(context, mColumnCount));
+            }
+            recyclerView.setAdapter(new OpenEventsRecyclerViewAdapter(PlaceholderContent.ITEMS));
+        }
+        return view;
     }
-}
+
+    }
