+++ EventAppTest2/app/src/main/res/layout/open_events_list.xml
