--- fragment_orgainzers_event.xml
@@ -1,99 +1,0 @@
-<androidx.cardview.widget.CardView
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_margin="8dp">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:padding="16dp">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
-            android:layout_width="match_parent"
-            android:layout_height="242dp">
-
-            <androidx.cardview.widget.CardView
-                android:id="@+id/OpenEventImage"
-                android:layout_width="380dp"
-                android:layout_height="160dp"
-                app:cardBackgroundColor="@color/white"
-                app:cardCornerRadius="5dp"
-
-
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent">
-
-                <androidx.cardview.widget.CardView
-                    android:layout_width="380dp"
-                    android:layout_height="160dp"
-                    android:layout_gravity="center"
-                    app:cardCornerRadius="5dp">
-
-                    <ImageView
-                        android:id="@+id/OpenUserImage"
-                        android:layout_width="390dp"
-                        android:layout_height="160dp"
-                        android:scaleType="centerCrop"
-                        android:src="@drawable/karansbackground" />
-
-                </androidx.cardview.widget.CardView>
-
-            </androidx.cardview.widget.CardView>
-
-
-            <!-- Username -->
-
-
-            <TextView
-                android:id="@+id/OpenEventTitle"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="64dp"
-                android:text="THE EVENT TITLE IS"
-                android:textSize="18sp"
-                android:textStyle="bold"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/OpenEventImage"
-                app:layout_constraintVertical_bias="0.0" />
-
-            <TextView
-                android:id="@+id/OpenEventlocation"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="44dp"
-                android:hint="ssssssssssssssss"
-                android:textSize="9sp"
-                android:textStyle="bold"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/OpenEventTitle"
-                app:layout_constraintVertical_bias="0.0" />
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/OpenEventButton"
-                android:layout_width="140dp"
-                android:layout_height="41dp"
-                android:layout_margin="16dp"
-                android:layout_marginBottom="16dp"
-                android:backgroundTint="@color/purple"
-                android:drawableEnd="@drawable/ic_buttonarrow"
-                android:drawablePadding="8dp"
-                android:gravity="center_vertical"
-                android:text="OPEN-EVENT"
-                android:textColor="@android:color/white"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintTop_toBottomOf="@+id/OpenEventlocation"
-                tools:layout_editor_absoluteX="0dp" />
-
-
-        </androidx.constraintlayout.widget.ConstraintLayout>
-
-    </LinearLayout>
-
-</androidx.cardview.widget.CardView>
